Resolve some compiler warnings so that I can build under dev mode
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@77770 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ cdr/cdr_adaptive_odbc.c
@@ -204,7 +204,7 @@ static int load_config(void)
 			if (entry->octetlen == 0)
 				entry->octetlen = entry->size;
 
-			ast_verb(10, "Found %s column with type %hd with len %ld, octetlen %ld, and numlen (%hd,%hd)\n", entry->name, entry->type, entry->size, entry->octetlen, entry->decimals, entry->radix);
+			ast_verb(10, "Found %s column with type %hd with len %ld, octetlen %ld, and numlen (%hd,%hd)\n", entry->name, entry->type, (long) entry->size, (long) entry->octetlen, entry->decimals, entry->radix);
 			/* Insert column info into column list */
 			AST_LIST_INSERT_TAIL(&(tableptr->columns), entry, list);
 			res = 0;
